# Conflicts:
#	packages/nomad/main.tf
#	packages/nomad/orchestrator.hcl
#	packages/nomad/template-manager.hcl
#	packages/orchestrator/internal/template/build/hash_index.go

+++ packages/shared/pkg/storage/storage_aws.go
@@ -65,6 +65,11 @@ func (a *AWSBucketStorageProvider) DeleteObjectsWithPrefix(ctx context.Context, 
 		objects = append(objects, types.ObjectIdentifier{Key: obj.Key})
 	}
 
+	// AWS S3 delete operation requires at least one object to delete.
+	if len(objects) == 0 {
+		return nil
+	}
+
 	_, err = a.client.DeleteObjects(
 		ctx, &s3.DeleteObjectsInput{
 			Bucket: aws.String(a.bucketName),
